[Feature] fix warnings caused by type

+++ cocos/renderer/backend/Buffer.h
@@ -9,11 +9,11 @@ CC_BACKEND_BEGIN
 class Buffer : public cocos2d::Ref
 {
 public:
-    virtual void updateData(void* data, size_t size) = 0;
-    virtual void updateSubData(void* data, size_t offset, size_t size) = 0;
+    virtual void updateData(void* data, unsigned int size) = 0;
+    virtual void updateSubData(void* data, unsigned int offset, unsigned int size) = 0;
     
 protected:
-    Buffer(size_t size, BufferType type, BufferUsage usage)
+    Buffer(unsigned int size, BufferType type, BufferUsage usage)
     : _size(size)
     , _type(type)
     , _usage(usage)
@@ -23,7 +23,7 @@ protected:
     
     BufferUsage _usage = BufferUsage::READ;
     BufferType _type = BufferType::VERTEX;
-    size_t _size = 0;
+    unsigned int _size = 0;
 };
 
 CC_BACKEND_END
