+++ cocos/renderer/backend/metal/BufferMTL.h
@@ -8,11 +8,11 @@ CC_BACKEND_BEGIN
 class BufferMTL : public Buffer
 {
 public:
-    BufferMTL(id<MTLDevice> mtlDevice, size_t size, BufferType type, BufferUsage usage);
+    BufferMTL(id<MTLDevice> mtlDevice, unsigned int size, BufferType type, BufferUsage usage);
     ~BufferMTL();
     
-    virtual void updateData(void* data, size_t size) override;
-    virtual void updateSubData(void* data, size_t offset, size_t size) override;
+    virtual void updateData(void* data, unsigned int size) override;
+    virtual void updateSubData(void* data, unsigned int offset, unsigned int size) override;
     
     inline id<MTLBuffer> getMTLBuffer() const { return _mtlBuffer; }
