+++ cocos/renderer/backend/metal/BufferMTL.mm
@@ -2,7 +2,7 @@
 
 CC_BACKEND_BEGIN
 
-BufferMTL::BufferMTL(id<MTLDevice> mtlDevice, size_t size, BufferType type, BufferUsage usage)
+BufferMTL::BufferMTL(id<MTLDevice> mtlDevice, unsigned int size, BufferType type, BufferUsage usage)
 : Buffer(size, type, usage)
 {
     _mtlBuffer = [mtlDevice newBufferWithLength:size options:MTLResourceStorageModeShared];
@@ -13,13 +13,13 @@ BufferMTL::~BufferMTL()
     [_mtlBuffer release];
 }
 
-void BufferMTL::updateData(void* data, size_t size)
+void BufferMTL::updateData(void* data, unsigned int size)
 {
     assert(size <= _size);
     memcpy((char*)_mtlBuffer.contents, data, size);
 }
 
-void BufferMTL::updateSubData(void* data, size_t offset, size_t size)
+void BufferMTL::updateSubData(void* data, unsigned int offset, unsigned int size)
 {
     assert(offset + size <= _size);
     memcpy((char*)_mtlBuffer.contents + offset, data, size);
